backend/requirements.txt
asgiref==3.4.1
crispy-bootstrap5==0.5
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
